in progress of migrating to typescript, fix server issue
this commit begins the process of migrating to typescript.
ALL of the types will be set to any for now to get the server up
and running. This commit also fixes a cors issue when connecting
from github. The server was defaulting to a port 5000 when
the server should actually be accessing the port defined in .env